move jwt to auth
