+++ livestream/auth/jwt_test.go
@@ -1,4 +1,4 @@
-package main
+package auth
 
 import (
 	"github.com/stretchr/testify/assert"
